+++ R/get_valid_gene_lists.R
@@ -16,25 +16,27 @@
 #' @importFrom data.table uniqueN :=
 get_valid_gene_lists <- function(ctd,
                                  list_names,
+                                 list_name_column,
                                  gene_data,
                                  annotLevel = 1,
                                  verbose = TRUE){
-  Gene <- count <- . <- Phenotype <- NULL;
+  Gene <- count <- . <- NULL;
 
   messager("Validating gene lists..",v=verbose)
   ctd_genes <- rownames(ctd[[annotLevel]]$specificity_quantiles)
   shared_genes <- intersect(unique(gene_data$Gene),ctd_genes)
-  gene_data <- gene_data[Phenotype %in% list_names,
+  #### Filter phenotypes #####
+  gene_data <- gene_data[get(list_name_column) %in% list_names,
                            ][Gene %in% shared_genes,
                              ]
   gene_counts <- gene_data[,.(count=data.table::uniqueN(Gene)),
-                           by=c("ID","Phenotype")]
-  validLists <- unique(gene_counts[count>=4,]$Phenotype)
-  gene_data <- gene_data[Phenotype %in% validLists,]
-  gene_lists <- lapply(stats::setNames(unique(gene_data$Phenotype),
-                         unique(gene_data$Phenotype)),
+                           by=c("HPO_ID",list_name_column)]
+  validLists <- unique(gene_counts[count>=4,][[list_name_column]])
+  gene_data <- gene_data[get(list_name_column) %in% validLists,]
+  gene_lists <- lapply(stats::setNames(unique(gene_data[[list_name_column]]),
+                                       unique(gene_data[[list_name_column]])),
          function(pheno_i){
-      unique(gene_data[Phenotype == pheno_i,]$Gene)
+      unique(gene_data[get(list_name_column) == pheno_i,]$Gene)
   })
   messager(formatC(length(gene_lists),big.mark = ","),"/",
            formatC(length(list_names),big.mark = ","),