+++ R/subset_phenos.R
@@ -19,7 +19,7 @@ subset_phenos <- function(cell_type,
                           q_threshold = 0.0005,
                           fold_threshold = 1,
                           verbose = TRUE) {
-  # templateR:::args2vars(subset_phenos)
+  # devoptera::args2vars(subset_phenos)
 
   #### Subset by q, fold_change, and celltype ####
   phenos <- subset_results(cell_type = cell_type,